graphics.scala reworked, the entity.scala, game.scala files have been created with some basic classes, removed unused code in main.scala.
The DisplayEntity and AnimatedEntity have fused in AnimatedEntity.
game.scala contains a object with basic code for the upcoming content.
Added several types of entities, first the generic class Entity from
which every subclass will inherit from, but also a SentientEntity class
for entities with characteristcs such as hp, constitution. The Player
class is an exemple of SentientEntity.